jetty 12

+++ src/main/java/net/whydah/admin/user/UserResource.java
@@ -1,5 +1,6 @@
 package net.whydah.admin.user;
 
+import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.*;
 import net.whydah.admin.errorhandling.AppException;
@@ -11,8 +12,7 @@ import net.whydah.sso.user.types.UserApplicationRoleEntry;
 import net.whydah.sso.user.types.UserIdentity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 
@@ -20,18 +20,38 @@ import java.util.List;
  * @author <a href="bard.lind@gmail.com">Bard Lind</a>
  */
 @Path("/{applicationtokenid}/{userTokenId}/user")
-@Controller
+@Component
 public class UserResource {
     private static final Logger log = LoggerFactory.getLogger(UserResource.class);
-    UserService userService;
 
-    @Context
+    private UserService userService;
+
     private Request request;
 
+    /**
+     * Default constructor for HK2
+     */
+    public UserResource() {
+        log.debug("Default constructor called by HK2");
+    }
 
-    @Autowired
+    /**
+     * Constructor with dependency injection.
+     * Using @Inject which works with both Spring and HK2
+     */
+    @Inject
     public UserResource(UserService userService) {
+        log.debug("Constructor injection called with service: {}", userService);
         this.userService = userService;
+    }
+
+    /**
+     * Sets the JAX-RS request context.
+     * This will be called by the JAX-RS runtime.
+     */
+    @Context
+    public void setRequest(Request request) {
+        this.request = request;
     }
 
     /**
@@ -43,38 +63,46 @@ public class UserResource {
      *
      * @param userJson xml representing a User
      * @return Application
-     * @throws AppException 
+     * @throws AppException
      */
     @POST
     @Path("/")
     @Consumes({MediaType.APPLICATION_JSON})
     @Produces({MediaType.APPLICATION_JSON})
-    public Response createUser(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
-                               String userJson, @Context Request request) throws AppException {
+    public Response createUser(@PathParam("applicationtokenid") String applicationTokenId,
+                               @PathParam("userTokenId") String userTokenId,
+                               String userJson,
+                               @Context Request request) throws AppException {
         log.trace("createUser is called with userJson={}", userJson);
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
+
         MediaType responseMediaType = findPreferredResponseMediaType();
 
         UserIdentity createdUser;
         String userResponse;
         UserAggregate userAggregate = null;
 
-
         createdUser = userService.createUser(applicationTokenId, userTokenId, userJson);
 
-
         if (createdUser != null) {
-        	userAggregate = UserAggregateMapper.fromUserAggregateNoIdentityJson(UserIdentityMapper.toJson(createdUser));
-        	return Response.ok(UserAggregateMapper.toJson(userAggregate)).build();
+            userAggregate = UserAggregateMapper.fromUserAggregateNoIdentityJson(UserIdentityMapper.toJson(createdUser));
+            return Response.ok(UserAggregateMapper.toJson(userAggregate)).build();
         } else {
-        	return Response.status(Response.Status.NO_CONTENT).build();
+            return Response.status(Response.Status.NO_CONTENT).build();
         }
-
     }
 
     @Deprecated //TODO merge with normal endpoint
     @POST
     @Path("/xml")
-    public Response createUserFromXml(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId, String userXml, @Context Request request) throws AppException {
+    public Response createUserFromXml(@PathParam("applicationtokenid") String applicationTokenId,
+                                      @PathParam("userTokenId") String userTokenId,
+                                      String userXml,
+                                      @Context Request request) throws AppException {
         return createUser(applicationTokenId, userTokenId, userXml, request);
     }
 
@@ -82,51 +110,76 @@ public class UserResource {
     @GET
     @Path("/{uid}")
     @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
-    public Response getUserIdentity(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
-                                    @PathParam("uid") String uid, @Context Request req) throws AppException {
+    public Response getUserIdentity(@PathParam("applicationtokenid") String applicationTokenId,
+                                    @PathParam("userTokenId") String userTokenId,
+                                    @PathParam("uid") String uid,
+                                    @Context Request req) throws AppException {
         MediaType mediaType = findPreferredResponseMediaType();
-        //MediaType responseMediaType = findPreferedResponseType(req);
         log.trace("getUserIdentity is called with uid={}. Preferred mediatype from client {}", uid, mediaType);
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
+
         String userResponse;
         UserIdentity userIdentity = null;
-
 
         userIdentity = userService.getUserIdentity(applicationTokenId, userTokenId, uid);
         userResponse = UserIdentityMapper.toJson(userIdentity);
         return Response.ok(userResponse).build();
-       
     }
 
     @PUT
     @Path("/{uid}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response updateUserIdentity(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
-                                       @PathParam("uid") String uid, String userIdentityJson) throws AppException {
+    public Response updateUserIdentity(@PathParam("applicationtokenid") String applicationTokenId,
+                                       @PathParam("userTokenId") String userTokenId,
+                                       @PathParam("uid") String uid,
+                                       String userIdentityJson) throws AppException {
         log.trace("updateUserIdentity: uid={}, userIdentityJson={}", uid, userIdentityJson);
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
+
         return userService.updateUserIdentity(applicationTokenId, userTokenId, uid, userIdentityJson);
     }
 
 
     @DELETE
     @Path("/{uid}")
-    public Response deleteUser(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
+    public Response deleteUser(@PathParam("applicationtokenid") String applicationTokenId,
+                               @PathParam("userTokenId") String userTokenId,
                                @PathParam("uid") String uid) throws AppException {
         log.info("deleteUser, uid={}, ", uid);
 
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
 
         userService.deleteUser(applicationTokenId, userTokenId, uid);
         return Response.status(Response.Status.NO_CONTENT).build();
-       
     }
 
     @POST
     @Path("/changePassword/{username}")
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.APPLICATION_XML)
-    public Response changePassword(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
-                                   @PathParam("username") String userName, String password) throws AppException {
+    public Response changePassword(@PathParam("applicationtokenid") String applicationTokenId,
+                                   @PathParam("userTokenId") String userTokenId,
+                                   @PathParam("username") String userName,
+                                   String password) throws AppException {
         log.trace("changePassword is called with username={}", userName);
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
+
         boolean isPasswordUpdated = false;
         isPasswordUpdated = userService.changePassword(applicationTokenId, userTokenId, userName, password);
         String passwdOk = "<passwordUpdated><username>" + userName + "</username><status>" + isPasswordUpdated + "</status></passwordUpdated>";
@@ -137,9 +190,16 @@ public class UserResource {
     @GET
     @Path("/{uid}/roles")
     @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
-    public Response getRoles(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId, @PathParam("uid") String uid) {
+    public Response getRoles(@PathParam("applicationtokenid") String applicationTokenId,
+                             @PathParam("userTokenId") String userTokenId,
+                             @PathParam("uid") String uid) {
         log.trace("getRoles, uid={}", uid);
         MediaType responseMediaType = findPreferredResponseMediaType();
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
 
         try {
             String body = null;
@@ -157,9 +217,18 @@ public class UserResource {
     }
 
     private MediaType findPreferredResponseMediaType() {
-        List<Variant> availableVariants = Variant.mediaTypes(MediaType.APPLICATION_JSON_TYPE, MediaType.APPLICATION_XML_TYPE).add().build();
+        if (request == null) {
+            log.warn("Request context is null, defaulting to JSON media type");
+            return MediaType.APPLICATION_JSON_TYPE;
+        }
+
+        List<Variant> availableVariants = Variant.mediaTypes(
+                MediaType.APPLICATION_JSON_TYPE,
+                MediaType.APPLICATION_XML_TYPE
+        ).add().build();
+
         Variant bestMatch = request.selectVariant(availableVariants);
-        return bestMatch.getMediaType();
+        return bestMatch != null ? bestMatch.getMediaType() : MediaType.APPLICATION_JSON_TYPE;
     }
 
 
@@ -176,39 +245,59 @@ public class UserResource {
     @Path("/{uid}/role")
     @Consumes({MediaType.APPLICATION_JSON})
     @Produces({MediaType.APPLICATION_JSON})
-    public Response addRole(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
-                            @PathParam("uid") String uid, String roleXmlOrJson) throws AppException {
+    public Response addRole(@PathParam("applicationtokenid") String applicationTokenId,
+                            @PathParam("userTokenId") String userTokenId,
+                            @PathParam("uid") String uid,
+                            String roleXmlOrJson) throws AppException {
         log.trace("addRole is called with uid={}, roleJson={}", uid, roleXmlOrJson);
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
 
         UserApplicationRoleEntry roleRequest = UserRoleMapper.fromJson(roleXmlOrJson);
         UserApplicationRoleEntry roleRepresentation = userService.addUserRole(applicationTokenId, userTokenId, uid, roleRequest);
         return Response.ok(roleRepresentation.toJson()).build();
-
     }
 
     @PUT
     @Path("/{uid}/role/{roleid}")
     @Consumes({MediaType.APPLICATION_JSON})
     @Produces({MediaType.APPLICATION_JSON})
-    public Response updateRole(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
-                            @PathParam("uid") String uid, @PathParam("roleid") String roleid, String roleJson) throws AppException {
-        log.trace("updateRole is called with uid={}, roleid={},roleJson={}", uid, roleid,roleJson);
+    public Response updateRole(@PathParam("applicationtokenid") String applicationTokenId,
+                               @PathParam("userTokenId") String userTokenId,
+                               @PathParam("uid") String uid,
+                               @PathParam("roleid") String roleid,
+                               String roleJson) throws AppException {
+        log.trace("updateRole is called with uid={}, roleid={},roleJson={}", uid, roleid, roleJson);
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
 
         UserApplicationRoleEntry roleRequest = UserRoleMapper.fromJson(roleJson);
         UserApplicationRoleEntry updatedRole = userService.updateUserRole(applicationTokenId, userTokenId, uid, roleRequest);
         return Response.ok(updatedRole.toJson()).build();
-        
     }
 
 
     @DELETE
     @Path("/{uid}/role/{roleid}")
-    public Response deleteRole(@PathParam("applicationtokenid") String applicationTokenId, @PathParam("userTokenId") String userTokenId,
-                               @PathParam("uid") String uid, @PathParam("roleid") String roleid) throws AppException {
+    public Response deleteRole(@PathParam("applicationtokenid") String applicationTokenId,
+                               @PathParam("userTokenId") String userTokenId,
+                               @PathParam("uid") String uid,
+                               @PathParam("roleid") String roleid) throws AppException {
         log.trace("deleteRole, uid={}, roleid={}", uid, roleid);
+
+        if (userService == null) {
+            log.error("UserService is null - dependency injection failed!");
+            return Response.serverError().entity("Service unavailable").build();
+        }
+
         userService.deleteUserRole(applicationTokenId, userTokenId, uid, roleid);
         return Response.status(Response.Status.NO_CONTENT).build();
-
     }
 
     @GET
@@ -218,6 +307,4 @@ public class UserResource {
     public Response ping() {
         return Response.ok("pong").build();
     }
-
-
 }
